Refactored Frontend and Filesystem. Frontend now displays a settings icon, a spot for weather data, and removed features not needed for the application(markers and polylines)
