+++ frontend/react-native-frontend/src/components/SafeScreen.js
@@ -1,0 +1,24 @@
+import React from 'react';
+import { View } from 'react-native';
+import { useSafeAreaInsets } from 'react-native-safe-area-context';
+
+export default function SafeScreen({ children, style }) {
+  const insets = useSafeAreaInsets();
+
+  return (
+    <View
+      style={[
+        {
+          paddingTop: insets.top,
+          paddingBottom: insets.bottom,
+          paddingLeft: insets.left,
+          paddingRight: insets.right,
+          flex: 1,
+        },
+        style,
+      ]}
+    >
+      {children}
+    </View>
+  );
+}
