+++ frontend/react-native-frontend/styles/themes.js
@@ -1,0 +1,7 @@
+// HEADER
+export const styles_header = {
+    height: 60,
+    borderRadius: 10,
+    margin: 7,
+    color: '#cccfd3ff',
+};
